@@ -15,13 +15,15 @@ ImentoreBase::Application.configure do
     # Log error messages when you accidentally call methods on nil
     # config.whiny_nils = true
   config.eager_load = false
+  config.log_level = :debug
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
   # Raise exceptions instead of rendering exception templates
-  config.action_dispatch.show_exceptions = false
+  # config.action_dispatch.show_exceptions = false
+  config.action_dispatch.show_exceptions = true
 
   # Disable request forgery protection in test environment
   config.action_controller.allow_forgery_protection    = false
@@ -37,5 +39,8 @@ ImentoreBase::Application.configure do
   # Print deprecation notices to the stderr
   config.active_support.deprecation = :stderr
 
-  config.action_mailer.default_url_options = { :host => 'localhost:3000' }
+  config.action_mailer.default_url_options = { :host => 'localhost' }
+
+  config.assets.digest = true
+  config.assets.debug = true
 end
