@@ -1,10 +1,15 @@
 # This migration comes from imentore_cms (originally 20120307114526)
 class AlterThemesAddDefault < ActiveRecord::Migration
+
   def up
-    add_column(:imentore_themes, :default, :boolean, default: false)
+    create_table :imentore_themes do |t|
+      t.string :name
+      t.references :store
+      t.timestamps
+    end
   end
 
   def down
-    remove_column(:imentore_themes, :default)
+    drop_table(:imentore_themes)
   end
 end
