@@ -5,11 +5,10 @@
 // or vendor/assets/javascripts of plugins, if any, can be referenced here using a relative path.
 //
 //
-//= require jquery/jquery-2.1.1.js
-//= require bootstrap-sprockets
-//= require metisMenu/jquery.metisMenu.js
-//= require pace/pace.min.js
-//= require peity/jquery.peity.min.js
-//= require slimscroll/jquery.slimscroll.min.js
-//= require inspinia.js
-
+// require jquery/jquery-2.1.1.js
+// require bootstrap-sprockets
+// require metisMenu/jquery.metisMenu.js
+// require pace/pace.min.js
+// require peity/jquery.peity.min.js
+// require slimscroll/jquery.slimscroll.min.js
+// require inspinia.js
