@@ -19,10 +19,7 @@ module Imentore
       scope :active, -> { where(active: true) }
 
       def self.install_store(store)
-        self.active.each do |admin_theme|
-          admin_theme.install_store(store)
-        end
-        store.themes.first.update_attribute(:active, true)
+        install_themes(store)
       end
 
       def install_stores
@@ -83,34 +80,69 @@ module Imentore
 
       private 
 
-      def theme_templates_store(store)
-        theme = Imentore::Theme.create(name: self.name, admin_imentore_theme_id: self.id, store: store, system: true)
-        theme.folders.create(name: "assets", store: store)
-        self.templates.each do |manager|
-          template = theme.templates.new
-          template.path = manager.path
-          # template.layout = manager.layout
-          template.kind = manager.kind
-          template.layout_id = manager.layout
-          template.default = true if manager.kind == "layout"
-          template.body = manager.body
-          template.header_view = ""
-          template.admin_imentore_template_id = manager.id
-          if template.save
-            if template.kind  == "layout"
-              @id = template.id
-            else
-              template.update_attribute(:layout_id, @id)
+      def self.install_themes(store)
+        # Remember, this runs on the current directory
+        base_directory = File.expand_path("#{Rails.root}/../imentore-manager/public/defaults/themes/")
+        init = YAML.load_file(File.expand_path("#{base_directory}/initializer.yml"))
+        init.keys.each do |themes|
+          next if File.directory?(File.expand_path("#{base_directory}/#{themes}"))
+          
+          theme_directory = themes.to_s.rjust(2,'0')
+          theme = Imentore::Theme.create(name: init[themes]["name"], store: store, system: true)
+          folder = theme.folders.create(name: "assets", store: store)
+
+          Dir.chdir File.expand_path("#{base_directory}/#{theme_directory}/assets")  
+          Dir.glob("*").each do |asset|
+              asset_file = Pathname.new("#{Dir.pwd}/#{asset}").open
+              folder.assets.create(file: asset_file, theme: theme)
+          end
+
+          ["layouts", "templates"].each do |template| 
+            Dir.chdir File.expand_path("#{base_directory}/#{theme_directory}/")  
+            Dir.glob("*").each do |file|
+              path = init[themes][template][file]
+              next if File.directory?(file) or File.exist?("#{Dir.pwd}/file") or path.nil?
+              body = File.open("#{base_directory}/#{theme_directory}/#{file}", "rb").read
+              if template == "layouts"
+                @layout = theme.templates.create(path: path, kind: template.singularize, default: true, body: body, header_view:"")
+              else  
+                theme.templates.create(path: path, kind: template.singularize, default: true, body: body, header_view:"", layout_id: @layout.id)
+              end
             end
           end
-          theme.templates.each do |template|
-            template.layout_id = theme.templates.layouts.first.id if template.kind == "template"
-          end
+          store.themes.first.update_attribute(:active, true)
+          return true
         end
-        self.assets.each do |admin_asset|
-          asset = theme.folders.first.assets.create(file: admin_asset.file, theme: theme)
-        end      
       end
+
+      # def theme_templates_store(store)
+      #   theme = Imentore::Theme.create(name: self.name, admin_imentore_theme_id: self.id, store: store, system: true)
+      #   theme.folders.create(name: "assets", store: store)
+      #   self.templates.each do |manager|
+      #     template = theme.templates.new
+      #     template.path = manager.path
+      #     # template.layout = manager.layout
+      #     template.kind = manager.kind
+      #     # template.layout_id = manager.layout
+      #     template.default = true if manager.kind == "layout"
+      #     template.body = manager.body
+      #     template.header_view = ""
+      #     template.admin_imentore_template_id = manager.id
+      #     if template.save
+      #       if template.kind  == "layout"
+      #         @id = template.id
+      #       else
+      #         template.update_attribute(:layout_id, @id)
+      #       end
+      #     end
+      #     theme.templates.each do |template|
+      #       template.layout_id = theme.templates.layouts.first.id if template.kind == "template"
+      #     end
+      #   end
+      #   self.assets.each do |admin_asset|
+      #     asset = theme.folders.first.assets.create(file: admin_asset.file, theme: theme)
+      #   end      
+      # end
 
     end
   end
