@@ -1,64 +1,92 @@
-# require "bundler/setup"
-require "pry"
-require "pry-nav"
-
-# Bundler.require(:default, :test)
-
-# ActionMailer::Base.delivery_method = :test
-# ActionMailer::Base.perform_deliveries = true
-# ActionMailer::Base.default_url_options[:host] = "test.com"
-
-# SHOUT! http://guides.rubyonrails.org/2_3_release_notes.html#quieter-backtraces
-Rails.backtrace_cleaner.remove_silencers!
-
-# Configure capybara for integration testing
-require "capybara/rails"
-require "capybara/rspec"
-Capybara.javascript_driver = :webkit
-
-require "database_cleaner"
-DatabaseCleaner.strategy = :transaction
-
-require "shoulda-matchers"
-
-require "imentore-testing/factories"
-
+# This file was generated by the `rails generate rspec:install` command. Conventionally, all
+# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
+# The generated `.rspec` file contains `--require spec_helper` which will cause
+# this file to always be loaded, without a need to explicitly require it in any
+# files.
+#
+# Given that it is always loaded, you are encouraged to keep this file as
+# light-weight as possible. Requiring heavyweight dependencies from this file
+# will add to the boot time of your test suite on EVERY test run, even for an
+# individual file that may not need all of that loaded. Instead, consider making
+# a separate helper file that requires the additional dependencies and performs
+# the additional setup, and require it from the spec files that actually need
+# it.
+#
+# The `.rspec` file also contains a few flags that are not defaults but that
+# users commonly want.
+#
+# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  config.include Imentore::Core::Engine.routes.url_helpers
-  config.include Devise::TestHelpers, type: :controller
-
-  config.before(:each) do
-    DatabaseCleaner.start
-    Capybara.reset_sessions!
-    # https://github.com/rspec/rspec-rails/pull/471
-    #
-    # Currently I need to set routes manually in before block in every test. I
-    # can't do it in RSpec config because before block of ControllerExampleGroup
-    # runs after before block in RSpec config and overwrites @route variable.
-    #
-    # @routes = Imentore::Core::Engine.routes
-    #
-    # http://stackoverflow.com/a/7967503
-    # http://gavinmorrice.com/blog/posts/22-how-to-test-routes-in-a-rails-3-1-mountable-engine
-    # http://www.builtfromsource.com/2011/09/21/testing-routes-with-rails-3-1-engines/
-    #
-    # use_route: doesn't seem to work in rails 3.2
-    # http://stackoverflow.com/a/5832908
-    # http://stackoverflow.com/a/7506094
-
-    Imentore.config.domain = "imentore.dev"
+  # rspec-expectations config goes here. You can use an alternate
+  # assertion/expectation library such as wrong or the stdlib/minitest
+  # assertions if you prefer.
+  config.expect_with :rspec do |expectations|
+    # This option will default to `true` in RSpec 4. It makes the `description`
+    # and `failure_message` of custom matchers include text for helper methods
+    # defined using `chain`, e.g.:
+    #     be_bigger_than(2).and_smaller_than(4).description
+    #     # => "be bigger than 2 and smaller than 4"
+    # ...rather than:
+    #     # => "be bigger than 2"
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
-  config.after(:each) do
-    DatabaseCleaner.clean
+  # rspec-mocks config goes here. You can use an alternate test double
+  # library (such as bogus or mocha) by changing the `mock_with` option here.
+  config.mock_with :rspec do |mocks|
+    # Prevents you from mocking or stubbing a method that does not exist on
+    # a real object. This is generally recommended, and will default to
+    # `true` in RSpec 4.
+    mocks.verify_partial_doubles = true
   end
 
-  # DatabaseCleaner will handle the transactions
-  # We'll use :truncation only for Cucumber and Capybara w/ js: true
-  config.use_transactional_fixtures = false
+# The settings below are suggested to provide a good initial experience
+# with RSpec, but feel free to customize to your heart's content.
+=begin
+  # These two settings work together to allow you to limit a spec run
+  # to individual examples or groups you care about by tagging them with
+  # `:focus` metadata. When nothing is tagged with `:focus`, all examples
+  # get run.
+  config.filter_run :focus
+  config.run_all_when_everything_filtered = true
 
-  # If true, the base class of anonymous controllers will be inferred
-  # automatically. This will be the default behavior in future versions of
-  # rspec-rails.
-  config.infer_base_class_for_anonymous_controllers = true
+  # Allows RSpec to persist some state between runs in order to support
+  # the `--only-failures` and `--next-failure` CLI options. We recommend
+  # you configure your source control system to ignore this file.
+  config.example_status_persistence_file_path = "spec/examples.txt"
+
+  # Limits the available syntax to the non-monkey patched syntax that is
+  # recommended. For more details, see:
+  #   - http://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/
+  #   - http://www.teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
+  #   - http://rspec.info/blog/2014/05/notable-changes-in-rspec-3/#zero-monkey-patching-mode
+  config.disable_monkey_patching!
+
+  # Many RSpec users commonly either run the entire suite or an individual
+  # file, and it's useful to allow more verbose output when running an
+  # individual spec file.
+  if config.files_to_run.one?
+    # Use the documentation formatter for detailed output,
+    # unless a formatter has already been configured
+    # (e.g. via a command-line flag).
+    config.default_formatter = 'doc'
+  end
+
+  # Print the 10 slowest examples and example groups at the
+  # end of the spec run, to help surface which specs are running
+  # particularly slow.
+  config.profile_examples = 10
+
+  # Run specs in random order to surface order dependencies. If you find an
+  # order dependency and want to debug it, you can fix the order by providing
+  # the seed, which is printed after each run.
+  #     --seed 1234
+  config.order = :random
+
+  # Seed global randomization in this process using the `--seed` CLI option.
+  # Setting this allows you to use `--seed` to deterministically reproduce
+  # test failures related to randomization by passing the same `--seed` value
+  # as the one that triggered the failure.
+  Kernel.srand config.seed
+=end
 end
